# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Avalio API",
    "description": "API do zarządzania zamówieniami i produktami",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "paths": {
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Pobierz wszystkie zamówienia",
        "description": "Zwraca listę wszystkich zamówień z ich produktami",
        "responses": {
          "200": {
            "description": "Lista zamówień",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Utwórz nowe zamówienie",
        "description": "Tworzy nowe zamówienie z listą produktów",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": {
                          "type": "integer",
                          "example": 1,
                          "description": "ID produktu (musi istnieć i być aktywny)"
                        },
                        "quantity": {
                          "type": "integer",
                          "example": 2,
                          "description": "Ilość (większa niż 0)"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Zamówienie utworzone pomyślnie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Błąd walidacji",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The items field is required."
                    },
                    "errors": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Nie udało się stworzyć zamówienia"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Pobierz szczegóły zamówienia",
        "description": "Zwraca szczegóły konkretnego zamówienia wraz z produktami",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID zamówienia",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Szczegóły zamówienia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Zamówienie nie znalezione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No query results for model [App\\Models\\Order] 123"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "CREATED",
            "enum": ["CREATED", "CONFIRMED"]
          },
          "total_price": {
            "type": "number",
            "format": "float",
            "example": 150.50
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-26 14:30:00"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "unit_price": {
            "type": "number",
            "format": "float",
            "example": 75.25
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Produkt XYZ"
          },
          "sku": {
            "type": "string",
            "example": "SKU-1234"
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Orders",
      "description": "Operacje na zamówieniach"
    }
  ]
}